// src/api/schemas/serviceAttachmentSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { ServiceAttachmentSchemaDescription } from './serviceAttachmentSchemaDescription';
import type { ServiceAttachmentSchemaTitle } from './serviceAttachmentSchemaTitle';

export interface ServiceAttachmentSchema {
  id: string;
  attachment_type: string;
  file_name: string;
  file_size: number;
  title?: ServiceAttachmentSchemaTitle;
  description?: ServiceAttachmentSchemaDescription;
  display_order: number;
  file_url: string;
}